backend/requirements.txt
# NetSale V6.0 Backend Dependencies
# 核心框架
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.18

# 数据库
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# 数据验证
pydantic==2.10.0
pydantic-settings==2.7.0

# 安全认证
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# CORS & WebSocket (内置 FastAPI)
# 工具
python-dotenv==1.0.1
